+++ src/main/java/eu/olli/cowmoonication/command/exception/ApiContactException.java
@@ -1,0 +1,7 @@
+package eu.olli.cowmoonication.command.exception;
+
+public class ApiContactException extends MooCommandException {
+    public ApiContactException(String api, String failedAction) {
+        super("Sorry, couldn't contact the " + api + " API and thus " + failedAction);
+    }
+}