+++ src/main/java/eu/olli/cowmoonication/command/exception/MooCommandException.java
@@ -1,0 +1,9 @@
+package eu.olli.cowmoonication.command.exception;
+
+import net.minecraft.command.CommandException;
+
+public class MooCommandException extends CommandException {
+    public MooCommandException(String msg) {
+        super("cowmoonication.commands.generic.exception", msg);
+    }
+}