+++ src/main/java/eu/olli/cowmoonication/util/Utils.java
@@ -11,6 +11,7 @@ import java.util.regex.Pattern;
 public final class Utils {
     public static final Pattern VALID_UUID_PATTERN = Pattern.compile("^(\\w{8})-(\\w{4})-(\\w{4})-(\\w{4})-(\\w{12})$");
     private static final Pattern VALID_USERNAME = Pattern.compile("^[\\w]{1,16}$");
+    private static final char[] LARGE_NUMBERS = new char[]{'k', 'm', 'b', 't'};
 
     private Utils() {
     }
@@ -51,5 +52,29 @@ public final class Utils {
                     DurationFormatUtils.formatDurationWords(duration, true, true),
                     dateFormatted);
         }
+    }
+
+    /**
+     * Formats a large number with abbreviations for each factor of a thousand (k, m, ...)
+     *
+     * @param number the number to format
+     * @return a String representing the number n formatted in a cool looking way.
+     * @see <a href="https://stackoverflow.com/a/4753866">Source</a>
+     */
+    public static String formatNumberWithAbbreviations(double number) {
+        return formatNumberWithAbbreviations(number, 0);
+    }
+
+    private static String formatNumberWithAbbreviations(double number, int iteration) {
+        @SuppressWarnings("IntegerDivisionInFloatingPointContext") double d = ((long) number / 100) / 10.0;
+        boolean isRound = (d * 10) % 10 == 0; //true if the decimal part is equal to 0 (then it's trimmed anyway)
+        // this determines the class, i.e. 'k', 'm' etc
+        // this decides whether to trim the decimals
+        // (int) d * 10 / 10 drops the decimal
+        return d < 1000 ? // this determines the class, i.e. 'k', 'm' etc
+                (d > 99.9 || isRound || d > 9.99 ? // this decides whether to trim the decimals
+                        (int) d * 10 / 10 : d + "" // (int) d * 10 / 10 drops the decimal
+                ) + "" + LARGE_NUMBERS[iteration]
+                : formatNumberWithAbbreviations(d, iteration + 1);
     }
 }